refactor: rename onboardingv2 to onboarding quiz
