+++ src/app/Scenes/Onboarding/OnboardingQuiz/Components/AnimatedFadingPill.tsx
