+++ src/app/Scenes/Onboarding/OnboardingQuiz/Components/AnimatedTooltip.tsx
