+++ src/app/Scenes/Onboarding/OnboardingQuiz/Components/ArtistListItem.tsx
