+++ src/app/Scenes/Onboarding/OnboardingQuiz/Components/OnboardingQuestionTemplate.tsx
@@ -3,7 +3,7 @@ import {
   OnboardingContextAction,
   State,
   useOnboardingContext,
-} from "app/Scenes/Onboarding/OnboardingV2/Hooks/useOnboardingContext"
+} from "app/Scenes/Onboarding/OnboardingQuiz/Hooks/useOnboardingContext"
 import { Box, Button, CheckCircleFillIcon, Flex, ProgressBar, Screen, Spacer, Text } from "palette"
 import { FC, useCallback, useState } from "react"
 import { AnimatedFadingPill, FADE_OUT_PILL_ANIMATION_DURATION } from "./AnimatedFadingPill"
