+++ src/app/Scenes/Onboarding/OnboardingQuiz/Components/PartnerListItem.tsx
