+++ src/app/Scenes/Onboarding/OnboardingQuiz/Hooks/__mocks__/useOnboardingTracking.ts
