+++ src/app/Scenes/Onboarding/OnboardingQuiz/OnboardingArtistsOnTheRise.tsx
