+++ src/app/Scenes/Onboarding/OnboardingQuiz/OnboardingArtistsOnTheRiseCollection.tsx
