+++ src/app/Scenes/Onboarding/OnboardingQuiz/OnboardingCuratedArtworks.tsx
