+++ src/app/Scenes/Onboarding/OnboardingQuiz/OnboardingFollowArtists.tsx
