+++ src/app/Scenes/Onboarding/OnboardingQuiz/OnboardingGene.tsx
@@ -1,8 +1,8 @@
 import { NavigationProp, useNavigation } from "@react-navigation/native"
 import { OnboardingGeneQuery } from "__generated__/OnboardingGeneQuery.graphql"
 import { FullScreenLoadingImage } from "app/Components/FullScreenLoadingImage"
-import { OnboardingResultsGrid } from "app/Scenes/Onboarding/OnboardingV2/Components/OnboardingResultsGrid"
-import { OnboardingNavigationStack } from "app/Scenes/Onboarding/OnboardingV2/OnboardingV2"
+import { OnboardingResultsGrid } from "app/Scenes/Onboarding/OnboardingQuiz/Components/OnboardingResultsGrid"
+import { OnboardingNavigationStack } from "app/Scenes/Onboarding/OnboardingQuiz/OnboardingQuiz"
 import { Button, Flex, Screen } from "palette"
 import { Suspense } from "react"
 import { graphql, useLazyLoadQuery } from "react-relay"
