+++ src/app/Scenes/Onboarding/OnboardingQuiz/OnboardingQuestionOne.tsx
@@ -2,8 +2,8 @@ import { useNavigation } from "@react-navigation/native"
 import {
   OPTION_NO_IM_JUST_STARTING_OUT,
   OPTION_YES_I_LOVE_COLLECTING_ART,
-} from "app/Scenes/Onboarding/OnboardingV2/config"
-import { useOnboardingTracking } from "app/Scenes/Onboarding/OnboardingV2/Hooks/useOnboardingTracking"
+} from "app/Scenes/Onboarding/OnboardingQuiz/config"
+import { useOnboardingTracking } from "app/Scenes/Onboarding/OnboardingQuiz/Hooks/useOnboardingTracking"
 import { useCallback } from "react"
 import { OnboardingQuestionTemplate } from "./Components/OnboardingQuestionTemplate"
 import { useOnboardingContext } from "./Hooks/useOnboardingContext"
