+++ src/app/Scenes/Onboarding/OnboardingQuiz/OnboardingSearchResults.tsx
@@ -2,7 +2,6 @@ import { useNavigation } from "@react-navigation/native"
 import { OnboardingSearchResults_viewer$key } from "__generated__/OnboardingSearchResults_viewer.graphql"
 import { OnboardingSearchResultsQuery } from "__generated__/OnboardingSearchResultsQuery.graphql"
 import { ArtistListItemPlaceholder } from "app/Components/ArtistListItem"
-import { useOnboardingTracking } from "app/Scenes/Onboarding/OnboardingV2/Hooks/useOnboardingTracking"
 import { extractNodes } from "app/utils/extractNodes"
 import { ProvidePlaceholderContext } from "app/utils/placeholders"
 import { times } from "lodash"
@@ -13,6 +12,7 @@ import { graphql, useLazyLoadQuery, usePaginationFragment } from "react-relay"
 import { ArtistListItemNew } from "./Components/ArtistListItem"
 import { PartnerListItem } from "./Components/PartnerListItem"
 import { useOnboardingContext } from "./Hooks/useOnboardingContext"
+import { useOnboardingTracking } from "./Hooks/useOnboardingTracking"
 
 interface OnboardingSearchResultsProps {
   entities: "ARTIST" | "PROFILE"
