+++ src/app/Scenes/Onboarding/OnboardingQuiz/OnboardingTopAuctionLots.tsx
