+++ src/app/Scenes/Onboarding/OnboardingQuiz/__tests__/OnboardingOrderedSet.tests.tsx
@@ -6,7 +6,7 @@ import { OnboardingOrderedSetScreen } from "../OnboardingOrderedSet"
 
 jest.unmock("react-relay")
 
-jest.mock("app/Scenes/Onboarding/OnboardingV2/Hooks/useOnboardingTracking")
+jest.mock("app/Scenes/Onboarding/OnboardingQuiz/Hooks/useOnboardingTracking")
 
 describe("OnboardingOrderedSet", () => {
   let env: ReturnType<typeof createMockEnvironment>
