+++ src/app/Scenes/Onboarding/OnboardingQuiz/__tests__/useOnboardingTracking.tests.ts
@@ -1,5 +1,5 @@
 import { ActionType, ContextModule, OwnerType } from "@artsy/cohesion"
-import { useOnboardingTracking } from "app/Scenes/Onboarding/OnboardingV2/Hooks/useOnboardingTracking"
+import { useOnboardingTracking } from "app/Scenes/Onboarding/OnboardingQuiz/Hooks/useOnboardingTracking"
 import { useTracking } from "react-tracking"
 
 jest.mock("react-tracking")
